Add editor's experimental Android support (and comment on web) to FAQ
This is master version of #5975 pull request.

Android was announced:
https://godotengine.org/article/dev-snapshot-godot-3-5-beta-3

although it isn't on any stable release/doesn't seem 100%
recommended yet, so marking it as experimental.

Commenting the lack of support of the Web exporting, that may be
unavailable for 4.0, and suggesting to use 3.X instead. This
seems to be planned for 4.1, but not promising anything just in
case.