Organise classes into a set of namespaces.
A complete and total set of namespaces have been designed to enable
better class discoverability by offering an alternative path for class
exploration where each namespace exists as a curated "Class Reference"
for the specific area of functionality it covers.

They also serve as an officially maintained set of namespaces to be
consumed by thirdparty Godot bindings so that classes can be organised
into sensible groups, rather than a single global namespace (which can
be an intimidating amount of symbols for a single
reference/module/package/namespace/library).

I decided to have namespaces defined in their own folder (rather than
inside the per-class xml files) so that they can/will be maintained
holistically. It may be worth adding a few scripts to support this, ie:

	* Detect missing classes.
	* Remove deleted classes.
	* Rename updated classes.

The "Class Reference" sections have been embedded directly into the
respective index files (would it be better suited to have them as their
own pages?).

Started by splitting the namespaces based on the functional teams:
https://godotengine.org/teams

I made some design calls around splitting core into engine, multicore
and math & followed up by integrating this with the tutorial sections
which are already nicely split across functional boundaries.

Core text functionality has been grouped together with translation.
Conceptually this works for me but the tutorial would need to be updated
to match (for this to make sense).

Also note that I've awkwardly placed the 'engine' namespace "Class
Reference" under the "Best Practices" page (I noticed that there doesn't
appear to be a general engine section/overview). I didn't want to orphan
it but perhaps a new 'engine/mainloop/scenetree' section could be added.

This is an implementation of #6137 and godotengine/godot-proposals#4543.
I chose not to make any changes to to godotengine/godot as I consider
the solution here to be primarily developer-experience &
documentation-focused and not a concern for any of the functional teams.
Even if class tags become a thing in Godot 5, I think having namespaces
defined here will still be valuable.