English improvements for content guidelines
This commit simplfies many sentences and rewords many sentences to make
them sound more natural. In some places, this commit rewords sentences
to make their structure more parallel. The goal of all of these changes
is to make the content guidelines easier to understand.

Before this commit, the paragraph at the bottom of “Guidelines and
principles” said pretty much the same thing that the paragraph at the
bottom of “Limiting cognitive load” said. This commit combines those two
paragraphs to  reduce redundancy.