[Styleguide] Removed setget
According to this docs page https://docs.godotengine.org/en/latest/tutorials/scripting/gdscript/gdscript_basics.html#properties (see note a bit below) setget is now default on all variables (at least if I didn't misunderstand it)

`setget` first of all doesn't compile in 4.0 and has been split into `get` and `set`
This is the solution to solve setget pretty much https://godotengine.org/qa/137477/how-to-use-godot-4-setget

However, even with this pull request, I ask to not merge because the code in the first link still won't compile, because set_active and set_state in variable declarations won't compile. While I could tweak them to work like the link above, I ask what is the ideal/recommended way by the core devs who changed the setget, because this is a **styling guide** and not a demo project where the goal is to just compile and run.