we should state early on that it requires SCons
Following the documentation step by step, it was missed that SCons is actually needed on the system to build and not only "uses" it.  Much later the docs clarify the requirements, but it's useful to state this differently at the very beginning.  This simple change would have saved me about 15 mins or so.