Specify Input Devices as an Advanced Setting
After reading this page, I was unable to find the "Input Devices/Pointing" section at first as it's hidden behind the "Advanced Settings" toggle in my version of the editor (version `v4.0.beta14.mono.official [28a24639c]`). This makes it explicit that you need to turn on Advanced settings to access the section.