Update 3d_gizmos.rst
As mentioned here https://github.com/godotengine/godot-docs/issues/6716
add/remove_node3d_gizmo_plugin is the wrong method name. But i'm not sure if it intentionally changed to add/remove_node_3d_gizmo_plugin.
Here are the adjustments anway in case it is intentional and going to stay that way.