Information about the need to download v. 4.72 of .Net
I had a problem with OmniSharp in VS code getting an "Error: The reference assemblies for .NETFramework,Version=v4.7.2 were not found. To resolve this, install the Developer Pack (SDK/Targeting Pack) for this framework version or retarget your application. You can download .NET Framework Developer Packs at...". I see that Godot should download it automatically (https://github.com/godotengine/godot/issues/40546) but It evidently stopped doing that, or sometimes fails to do that. That's why I'm adding this information for anyone, so that when they will get the same error and get flustrated, then they won't quit because of it.