clarification of connecting signals via code
The previous example for "connecting a signal via code" used the editor extensively. 
This version uses only code to spawn a timer and connect the relevant signal, and is a concise, standalone example of how it's done.
I believe that not using the editor at all is more appropriate for this section.