fix: 3d tutorial squash code
The code block was indented 5 spaces and not 4. This caused it to get pasted in as one tab with one extra space in the script editor.

This of course broke indentation and made the syntax highlighter have show an error on the whole block since it had the wrong indentation.