Update code to 4.0+ for screenshotting the viewport tutorial
The viewport tutorial:
https://docs.godotengine.org/en/latest/tutorials/rendering/viewports.html

did not work for Godot 4- as instead of get_data() from
Texture one should use get_image() from Texture2D.

Also setting a texture didn't work as documented, fixed too.
Tested by using the code on a simple scene with a sprite.

This should be backported to 4.0 branch, but not previous ones.

Fixes #7162