gdexample.gdextension file not recognized
Using Windows 10 and Godot 4.0.2 Mono

The gdexample.gdextension file used in the example project is not being recognized by my version of Godot. I don't know the exact reason but I'm guessing it is because it doesn't specify whether or not it is supposed to be a debug or release file.

My changed .gdextension file is the same as the one that is available under godot-cpp\test\demo\example.gdextension. With this one it worked with no problems.