Update submitting_to_assetlib.rst
Added asset file structure (addons/AssetName) to reccomendations. Assets on the Asset Library are inconsistent in file structure. This is often problematic, as the LICENSE file is typically placed inside of res://, which will usually either get overwritten or ignored. 
While the submission guidelines heavily imply this is how you should structure your assets, it was never explicitly stated.