Removed outdated and confuse statements from c# basics
The _preliminary benchmarks_ were done with a tool that was for version 3.X and is no longer in active development. So it doesn't make sense to keep outdated information on a tutorial page.

And even if it made sense, stating that one language can be 4x faster than the other in specific cases, and saying nothing about what those cases are doesn't help much.

The reason for the edit is the frustration of having multiple people come over time on the forum and discord thinking that by simply using C# they would be writing a more performatic game.