Custom editor icon section's command wrong or out of date.
In the "*.gdextension" code block, the [Icon] tag does not work in 4.1.1.stable version, and it does not match the naming conventions of other tags. I test it, [icons] tag is worked well. In addition, I added a blank lines between the tag and the content to make it more consistent with other "*.gdextension" code block formatting.

Update with capnm's suggestion. Thank you all for your attention and reply!