Update to the 2D tutorial
Issue #41283 is currently disrupting the flow of the official 2D tutorial for folks with external script editors.

This change is to update the tutorial to acknowledge this bug and provide newcomers with clearer guidelines on how to complete the tutorial despite the bug.

It also clarifies a potential point of confusion about how the editor communicates connection between signals and script functions.