Update Instantiate menu in 03.player_movement_code.rst
This commit updates the instructions for instantiating a new child
scene in the getting started first 3d game tutorial so that the
menu option matches the language used in version 4.1.1.

The associated image is updated to match and is updated to the
preferred `webp` format.

The "Instance Child Scene" menu now reads "Instantiate Child Scene".