Update complying_with_licenses.rst
Moved the tip regarding how to get license information in the Engine singleton to after the list of inclusion methods, as it read oddly saying "These are the most common approaches" and then the reader having to scroll quite a bit to see the list of approaches. Also removed engine contributor methods, as seemed unnecessary and potentially confusing.

On the other hand, moved the tip about custom builds up higher, to put it next to, and gave it equal prominence with, the info about default templates.

Finally, some minor wording changes, for grammar, for clarity, or to remove redundancy.