Clarify explanation of property initialization.
The previous text implied (at least to me) that there was something
special in Godot called an "init assignment value", and that there was
special logic determining which kind of assignment was used.  In
reality, nothing special is happening: the property is initialized, then
the constructor may set the value, then exports may as well.

I also harmonized the Godot and C# code samples, rather than expecting
the user to read both in order.

Frankly I almost feel like this would be clearer if the entire section
were simply deleted, but I wasn't sure if that would be accepted.