GetCollider() Not Working Following Tutorial With v4.1
When following the tutorial GetCollider() seems to act like a pop function and cannot be called multiple times, I changed it to only be called once and it worked. I put my example working code for GDScript but I'm new and not sure if that is a proper way to fix it.