Removed the following line of code from gdscript:
```
            if collision.get_collider() == null:
                continue
```

This line of code is not needed and is confusing to the reader. When the
player collides with the ground, the collision object will be the
ground itself. The code above is checking if the collision object is
null, which it never will be.

I spent some time trying to figure out why the floor (a StaticBody3d
node) would return null when colliding with the player. After adding
some logging I discovered that collision.get_collider() was returning an
object when colliding with the floor and not null and this if statement
was never being executed!

This line of code is actually not in the c# version of the script.