Update "The Main Game Scene" for clarity
Many people have reported having issues connecting HUD's start_game and the main node's new_game() in the next step of this tutorial. This occurred for me because the documentation states that "new_game()" should be removed from the _ready() function, but does not include instructions to add "pass" back (and pass is not left in the example code when "new_game()" is added to the function, so the initial one will not remain). Having the function declaration without "pass" breaks all the following functions, preventing new_game() (or any Main functions) from being connected.