Update viewports.rst
Changed the order in which the Cameras section is introduced, and added necessary information to the provided example. The behavior only occurs as specified when SubViewport has its own World3D, or a World3D assigned to it; it does not exclude rendering of MeshB from the root Viewport if it does not.