Reduce Sphinx build times by using multiple jobs
The default number of sphinx jobs is increased from 1 to 4, and the
number of jobs (N) can be set on the command line with

  make SPHINXJOBS=N

According to Steam's latest survey, over 90% of gamers (potential game
developers) had access to at least 4 CPU cores in December 2023, so
that's a sensible default.
Source: https://web.archive.org/web/20240127204540/https://store.steampowered.com/hwsurvey/cpus/

This can drastically reduce the time needed for a fresh rebuild, which
is especially important for new contributors. For my informal tests, I
used a Ryzen 5 3600XT PC, with a NVME SSD runing Linux. Here are my
results:

- With 1 job (make SPHINXJOBS=1), it takes about 25 minutes to build.
- With 12 (make SPHINXJOBS=12), I saw a 4x speed-up, bringing my build
  time down to just 6 minutes.
- With the default value of 4 jobs (make), I saw a 2x increase, with a
  build taking only 9 minutes.