Update signals.rst
I am going through the tutorials to learn godot.

Upon hitting this tutorial I had to find where my issue was specifically on halting the processing of the 2d sprite.

Upon finding where c# was using functions for _on_button_pressed() with godot, the example used "OnButtonPressed()". Fixing it to the exact name property helped as I am finding that godot is case sensitive and wants the exact spelling just like the default godot programming. 

From that instance learning that I am commiting fixes to the timer function for C# as well.

It should be noted for the user if they need the exact name and casing for C#, (or other languages for that matter they can go into the node properties of an object, right click the connection, click edit, and then copy and paste the function name to avoid syntax errors that can break the program.