More accurately describes what a Transform3D is
As pointed out in Issue #8904, a Transform3D is a 3x4 transformation matrix, not a 4x3. While that issue does point out other areas of concern on this specific page, this update is simply to fix that error. The change can be found on line 35.

(cherry picked from commit 9fdfe110044b050046a707a1b82ddbb5bb24babd)