I started to create a new page for the introduction section of the documentation. I copied introduction_to_godot.rst to external_resources.rst. external_resources.rst needs to be modified.
I would add a page, external_resources, that provides links to step-by-step tutorials as a marketing device that generates awareness of Godot, gets people to use Godot and gets people to make recommendations on what Godot can do to be more useful to them.

The intended audience is potential new users. These are game developers who are considering using Godot and need something to help them make the decision. I propose to provide links to video tutorials, so that these potential new users can explore the ins-and-outs of Godot so that they can make an informed decision. A secondary audience are people who want to provide how-to step-by-step tutorials. This will provide a place for them to put links to their tutorials.

When I was first getting started with Godot, I had trouble finding step-by-step tutorials. This is not the point of Godot’s documentation. For this reason, I think adding another page to Godot’s documentation that shows a list of external links for learning Godot would help. Godot’s documentation wasn’t intended to include step-by-step tutorials or how-to guides. I find those very useful, so I would add a page to the documentation that links external resources for those how-to-guides and step-by-step tutorials.