Update compiling_with_script_encryption_key.rst
Adding in the compiling a PCK with encryption. It may be necessary to note, that when adding the env var to Windows with command prompt, terminal, or Powershell, it might require being ran as admin. Not doing this can lead the the engine unable to unencrypt the PCK at runtime.