Update the getting started 3d tutorial for godot 4.x
I followed the entire 3d tutorial with godot 4.x and the changes to the
docs repo were small. The largest change was needing to replace the
outdated links to the old starter project with assets. Since there was
no longer a non-complete starter project, I made a PR to
godot-docs-project-starters. That would need to be reviewed and merged
first before the link to the new starter project zip file would be
valid.

New starter/assets project PR:
https://github.com/godotengine/godot-docs-project-starters/pull/9