Update JetBrains Rider documentation (#10025)
Revised JetBrains Rider setup steps and clarified debug configurations. Replaced outdated .webp images with new .png files for newer and clearer screenshots.

Add a note clarifying that this documentation is intended for contributing to the game engine, not for using JetBrains Rider as a C# or GDScript editor. Include reference to the C# guide for configuring an external editor.

Fixed duplicate link based on the Sphinx format. This should fix the warning/error on the page.

---------

Co-authored-by: Max Hilbrunner <mhilbrunner@users.noreply.github.com>