Pre-release commit storm
Before the release, a lot of testing was done, including the ci/cd pipeline. Some of this could
only be done in the live environment, which lead to many small commits with bad descriptions

Also includes many small fixes in urls and textual changes. Most of it inside migrate.py