Add new Renderers page
Adds a new page which gives a summary of the three renderers. This page briefly
describes the renderers, gives some guidance on how to choose a renderer. It also
contains a comparison table between the renderers.
Adds an entry in the FAQ which links to this new page.
Adds a note about subsurface scattering limitations to the Standard Material page.
Adjusts some other pages which had incorrect information.