Add note about LightmapGI only baking nodes under its parent
`LightmapGI`'s documentation currently doesn't mention the fact it only bakes nodes under its parent.
This can make the user think there is something wrong with their scene setup or 3D models, as it refuses to bake when the user's models / world isn't under the same parent as the `LightmapGI`.