Add info about layer mask exporting to physics_introduction.rst (#10242)
* Update physics_introduction.rst

The docs page for raycasting provides a link to the physics introduction page for how to set a collision mask on a raycast created in code. The physics introduction page does not mention export annotations for exporting a collision mask. This pull request adds a brief bit about exporting layer masks, and links to the export annotation docs.

---------

Co-authored-by: tetrapod <145553014+tetrapod00@users.noreply.github.com>