add flex layout for left nav container
this commit makes the left nav a flex container. by doing that, it also removes the fixed calculation that happens in custom.js. that manual calculation makes parts of the left nav non-scrollable in e.g. "stable" mode of the docs. with this commit, any changes to the top portion of the left nav's height will not matter as the bottom part will flex on the available height automatically.