Add GitHub Action to automatically create cherrypick PRs (#10361)
* Add GitHub Action to automatically create cherrypick PRs

Uses peter-evans/create-pull-request and actions/checkout,
both of which are already dependencies.

When merging a PR with a cherrypick label, which only contains
a single commit, attempt to cherrypick the single commit into
a new branch, then create a PR.

Fails if there are multiple commits (to catch squash and merge).
Fails if the cherrypick does not apply cleanly.

Does not attempt to merge the new PR.
Does not attempt to update stable.

* Handle squash and merge

* Apply suggestions from code review

Co-authored-by: Thaddeus Crews <repiteo@outlook.com>

---------

Co-authored-by: Thaddeus Crews <repiteo@outlook.com>