Point mapping contributions to Godot in Controllers, gamepads and joysticks
GameControllerDB's upstream does not accept contributions from Godot
since it does not use SDL. Users can still create custom mappings,
but they should be submitted to Godot's controller database instead
of the SDL one.

This also adds a reference to `godotcontrollerdb.txt` which was missing
in the first place. This file is used as a complement to `gamecontrollerdb.txt`.