Add note on pressing F6 for current scene vs. F5 for default scene
When following the tutorial instructions I reached the point of running the scene to test the button that starts and stops movement of the Godot icon.  After following the instructions explicitly, you will have the node_2d scene selected, but you are currently instructed to "run the scene".  If you choose to do so by pressing F5 or the play button in the UI, the default scene will run, which does not contain the button.

This change will add a note that explains the difference in behavior between F5 and F6 to help get people over the stumble and present a learning opportunity.

This issue has been encountered by others as well, as seen here:
https://www.reddit.com/r/godot/comments/16regpg/godot_411_button_not_showing_up/