Use global pos for mob spawn location
Fixes #6372 by correcting the spawn location of creeps,
accounting for the MobPath's position in the main scene,
which is 240, 450 (center of the rectangular path) and not 0, 0.

Without this change, creeps spawn from an offset position around
the origin 0,0, resulting in creeps appearing in the middle of the
screen when they spawn from the right edge, killing the player without
a chance for them to dodge.

Using global_position ensures creeps spawn on actual screen
edges using the world position of the MobSpawnLocation instead
of its relative position to its parent MobPath.