Help people create MobPath as a child of Main
Better steer people to create MobPath as a child of Main, so that they
don't accidentally create it as a child of StartPosition (the previously
selected item), which has a transform.

If they happen to re-parent to Main after creating the Path2D as a child
of StartPosition, they end up with a transform on the Path2D, which
breaks the expected spawn point of mobs.

See https://github.com/godotengine/godot-docs/pull/10698#issuecomment-2664002759.

Closes #6372, closes #10698.