Remove warning on build caused by non-existent pygment
At #10857 a gitignore codeblock was added. Addition seems correct,
but a pygment formatting was defined as:

.. code-block:: gitignore

No such a pygment exists, and although there is an aditional
package that could parse git-related files, we chose to
not add an extra package dependency for building the docs.

Fallback to unixconfig instead.

Closes #10989