Rewrote parts distance alpha section for clarity
"This is the most effect" doesn't make much sense, and saying it's the most effective doesn't make it clear on what it's most effective in doing.
Clarified that the pixel dither uses a 4x4 Bayer matrix. My saying that is based on simply looking at the example image up close and comparing it to the Wikipedia image example, so I'm pretty sure that's what it's using but if someone knows bettter then I'll change it.