Add section about angular velocity to using_transforms.rst
How to calculate angular velocity doesn't seem to be anywhere in the online doc, and it is very easy to incorrectly assume that they are Euler angles, which happened to me and was reported by others.
Since this is where quaternions and their importance are explained, it seems a good place to explain how to obtain correct angular velocity from quaternions.

This is a subject I had considerable difficulties with, and could only solve it thanks to the help of others. Notably the code here is simplified from code helpfully shared by Norgorim (any error is however mine).
Hence the need for an explanation in the documentation, and this appears to me as at least the necessary minimum.