Add JavaScript-based redirects to supplement Read the Docs redirects
Read the Docs has an upper limit on the number of redirects that can
apply. This limit is exceeded by the current `redirects.csv`,
which means many redirects don't apply correctly.

While this JavaScript-based solution is not as seamless from the
user's perspective, it has no limit on how many redirects can be defined.