Fix error in shader_functions abbr popup
When hovering over vec_uint_type the popup claims that it can accept a float instead of a uint. This pr fixes this error. The top of the page states correctly that it accepts a uint, uvec2, uvec3, or uvec4.

(cherry picked from commit f2023307bcac6e40192cb1d0309da7024505b7df)